trackerCell file refactoring
